.screenrc: Explicit setup of hardstatus line
The line was varying a lot due to different settings from different
distributions, so we override them.

We unfortunately now start with an empty %h; this should probably be
fixed by running `xtitle` on the start of a new interactive shell.
(This would also fix having the wrong title bar on ssh'ing into
another host.) In the meantime, as a warning of sorts, we display
"Gnu Screen" when there is no %h.