.bash_profile.inb2: Set BASH_ENV=~/.bashrc
Full explanation of why we do this at all is in the file.

We have this in cjs1 instead of cjs0 because other folks might have a
`.bashrc` that doesn't avoid doing bad things in non-interactive
environments (though this can't be all that likely, since it would
affect them every time they `ssh server some command`).