.config/htop/htoprc: Move meters for better viewing
All CPU meters go on right side, with overall average at top separated
from individual CPUs below by the load average. Mem and swap stay on
the left side but also move to the top for faster finding with the eye.

As it turns out, this doesn't change the header length at all.