.inputrc: More sophisticated bindings, Ctrl-L fix, add re-read
This now documents how we set separate bindings for vi-insert and
vi-command mode. (Previously we were setting only for insert mode
without realizing this.)

`Ctrl-l` is changed in both modes to merely redraw the current line
without clearing the screen. (This can also be achieved by preceding
`Ctrl-l` with a count, but this is easier.)

`Esc-r` in both modes re-reads the `.inputrc` file.

`Ctrl-o Ctrl-i d` in insert mode now inserts the string `$(i-date)`
(just like my binding in vim). I would prefer it actually runs the
command and inserts the value, but there seems to be no way to
configure that in this file. (Even using `bind -x` it seems to
immediately run the command as if typed at the command line, rather
than eating the output value and inserting it into the current command
line.) So this serves perhaps more as a demo, though it may be useful.