.config/htop: Improve top-of-page meters yet again
Ok, now on the right side we have just the CPU graph and we've moved
memory and swap bars over there as well. Total, 6 lines.

We can move average CPU over to the less-viewed left because we don't
use it nearly as much now that we have the graph. (In fact, we may not
need it at all, really.) All individual CPUs stay at left, too, making
it one line shorter than the right on a 4-core system, one line longer
on an 8-core system, and of course still hugely long on a 28-core system.