.bashrc: Put color back in prompt
The issue before was that `\[` isn't interpreted by readline; it's
interpreted only in some circumstances by the shell where it generates
`\001` and `\002`. Using those directly fixes our problem.

We make the prompt bold blue to make it a bit easier to see in the
scrollback; perhaps this will mean we don't need to use sp() so often
to make it clear where output from a particular command starts.