.clusterssh/config: Just 'xterm' instead of 'choose xterm'
It seems that since I last used this (4.02, now 4.08) one cannot
give arguments on the `terminal` property, and

    terminal=choose
    terminal_args=xterm

gives complaints about bad options from choose because it doesn't like
the xterm args for whatever reason. I'm kinda thinking long term that
I may not really need `choose` anyway, so for the moment I'll just
lock this to xterm, which is always available everywhere.