.bashrc __cjs_prompt(): Re-order colors to put normal mode last
`declare -l` prints the `$'...'` definiations as the raw control
characters in single quotes and the terminal interprets these
(correctly) as control sequences and so changes the colors. This
causes everything after those lines to come out in whatever
combination of colors happen to be the last set. Put `none` at the end
so that the last set is normal mode and everything after it now comes
out normally.