.bashrc: Better and faster username checking for prompt
We no longer use the `id` command because we were running it once,
sometimes twice, per prompt. This new method is a bit less reliable if
not root, but it's not likely to fail.

We now use the read-only $EUID Bash variable to check if we're root.
This can't be spoofed, so we're fine here.

If we're not root, we check our name against $USER, which is not a
Bash variable at all (it's set at login, I guess) which could be
spoofed, but who would do this and why? (Plus it would probably break
other stuff.)

Also extend our name check to cover 'csamp*' as well.