.bashrc: Unalias vi, vim etc. before defining functions
Aliases are, weirdly, expanded (but then ignored) when defining
functions. E.g., if `f` is an alias for `x y`, `f() { true; }` will be
expanded to `x y() { true; }` and produce an error. (However, if the
alias had been one word, no error and `f`, not the contents of the
alias, would be bound.)

In some cases, such as during upgrades or maybe just as an accident,
we're running in an environment with old aliases defined, so ensure
that we clear them out before defining our functions both to avoid
the above error and to avoid those aliases overriding our functions.
(Though do we really want to override them?)