.bashrc: Add `cd()` to set CWD to real dir after dereferencing symlinks
By default, Bash tries to pretend that, when you `cd` through a
symlink, that symlink really is a directory. (Thus, `cd ..` will
change to the parent of the symlink, rather than the parent of the
actual directory you're in.)

I prefer to be where I really am, so this figures out the actual dir
the symlink points to and changes to that, thus setting CWD to the
dereferenced symlink. It's careful to preserve `$OLDCWD` so that
`cd -` still works properly.

If a symlink was dereferenced, the actual directory to which we
changed is printed.