.gitconfig: Change tag and stash colours to be more consistent
Tags become cyan because that's closer to the blue and green we use
for branches and remote branches. Now they don't pop out so annoyingly
as when they where magenta.

The stash was popping out too much in listings for my taste; changing
it from cyan to bold yellow makes it much better. (While the colour is
now quite subtle, stashes are also highlighted by the branch and merge
commit they add to a graph listing.)