.gitconfig: Set "commit 0xa1b2c3" in logs to bold
It wasn't obvious how to do this until I asked on StackOverflow
(below); it turns out to be a setting under color.diff even though it
doesn't appear to be used by `git diff` at all, just by `git log`.

Bold black is a compromise between not being too loud in oneline
listings and making it easier to find the start of a commit in `git
log --patch`. The latter is still not as easy as I'd like it to be,
but while magenta was great for that, it's annoying when there's a lot
of it on the screen (in the oneline logs).

https://stackoverflow.com/a/50147068/107294