.lesskey: Attempt to emulate my gy/ge "small scroll" from vim
This doesn't work very well because one can't specify a count, so
you're limited to just 6-line scrolls. But at least this will do
something slightly similar to vim, even if it's not perfect.

Probably I should just get used to using counts on ^U/^D, since
that works the same in both. Then I could maybe map gu/gd to reset
the count to scroll a half screen, but then I'd need to find a new
home for the current nohlsearch on `gu`.