.ipython: Replace standard banner with custom, no blank line between commands
The standard banner isn't too bad in IPython 6 (3 lines), but is
rather annoying in IPython 2 (8 lines). Anyway, just turn it off and
add a startup file that prints our own one-line banner. (We may even
want a way to disable that at some point, but we'll see.)

Also set `separate_in` to nothing rather than a newline to get the
same effect as --nosep: no blank line between the output and the next
prompt.