bin/connect-proxy: setup stream connection via HTTP proxy CONNECT
This is basically just a wrapper around netcat that gets proxy
information from the `*_proxy` variables in the environment.

It currently supports only HTTP protocol to connect to the proxy,
but can easily be extended to other protocols when necessary.

We add a `pssh.github.com.` host to `.ssh/config` to demonstrate
how this can be used.