.vim/spell/en.utf-8.add: Personal dictionary moved from old repo
This is the uncompiled dictionary; I'm no longer committing the
compiled (`.add.spl`) file. Instead we have `dh/setup` compile it
using a bit of clever vim command-line trickery.

Unfortunately, there seems no easy way of getting vim to print out
error messages should :mkspell fail, but not print out the noise if
it's successful. So we just use `-e -s` to keep it quite quiet and
print a notice if it exits with a non-zero exit code.