.bashrc: Extract function to copy completion and pre-load completion
This fixes the bug where one had to try to complete the original
command in order to get the completion code loaded before completion
would work on the alias of the command.

This may share some code in common with the `git-tools` function to
copy completion. But it may not; I'm not sure if `git` is using
standard bash-completion conventions for autoload.