.bashrc: Set up command-line completion for pip and pip3
If we have pip3 we use it to set up both `pip3` and `pip` because,
even if the latter isn't available by default, it becomes available
when we activate a virtualenv. We have to change the completed command
from `pip` to `pip3` in older versions of pip and from `pip3` to `pip`
in newer versions of pip.