.bashrc,generate-completions: Generate completion for pip/pip3
Setting up the completion for pip/pip3 in .bashrc was taking from 0.8
to 1.8 seconds, which was significantly slowing opening new terminals.

We fix this by extracting the setup from .bashrc and moving it to
share/bash-completion/generate-completions, where it generates a
completions/pip file that bashrc then sources in the normal way it
sources everything in that directory. Rather than regenerating the
completion Bash code every time, we generate that file only if it
doesn't exist or is more than 24 hours old, thus avoiding generation
most of the time.

.bashrc is still slower than I'd care for, close to a second on my
basic Pentium system, so this code may change or be moved elsewhere
depending on what other things I find slowing stuff down.