share/.../generate-completions: Fail silently on very old pip
Some systems (e.g., Ubuntu 14.04) have a very old system pip without
`--disable-pip-version-check`. In this case we just generate no
completion because that's the easiest thing to do. If we actually want
to use pip enough on such a system that we'd want completion, we'd
also want to install our own user-local version of pip which would be
new enough to have this option.