.bashrc find(): Warn when more than one arg and last is a dir
In /etc/, `findf systemd cups` may look like it should find all files
under `systemd/` with `cups` in the name, but actually, because `cups`
is a directory, it finds all files with any name under both dirs. This
can be confusing.

Now, when there's more than one argument, indicating the possibility
that the user was trying to specify a name fragment, and the last arg
is a dir, we warn about this. We also pause for a second after the
warning in case there's a lot out output that would cause it to scroll
off the top of the terminal window. It's hacky, but this is only used
interactively.