.bashrc runvirtualenvcmd: Simplify completion function
virtualenvwrapper_load returns instantly if virtualenvwrapper is
already loaded, so there's no harm in calling it every time the
completion function is called rather than having special setup on the
first run to avoid calling it after that.

We also set the completion function to our own function, even though
all that does is directly call the _virtualenvs function with the same
arguments, so that we can later extend this function to complete
commands as well as virtualenv names.