.bashrc runvirtualenvcmd completion: Complete command name after env name
This now correctly completes the first parameter (environment name)
and second parameter (executable filename in the environment's bin/)
but does not yet complete arguments to the command. We should be able
to do this by using that command's completion if it exists outside the
virtual environment.

This also adds a comp_log function that logs completion information
for use when debugging completion commands.