dot/bashrc __runvirtualenvcmd_complete(): Fix when wrapper not available
If virtualenvwrapper is not installed, we now just return an empty
COMPREPLY and suppress any error messages.

It would be nice to know that it's not installed, but printing
anything during completion messes up the screen, which is more
annoying than not knowing if virtualenvwrapper is installed and there
are no envs or if it's not installed.