.mame/mame.ini: Basic MAME/MESS configuration
- Run in a moderate size window instead of full-screen.
- Don't capture the mouse.
- Set path for ROMs. (This can be a symlink.)

The uimodekey, which lets TAB bring up the MAME menus and various
other keys do their thing (resetting the machine and whatnot), is
INSERT on Linux by default, not Scroll Lock, despite what all the docs
and even the on-screen prompts say. We import this setting here just
to document it, though I'd like to look for a different key.