.bashrc: Move WORKON_HOME from ~/co/py-virtualenv to ~/.pyvirtenv
It's not really a checkout, and not following the current cpd
configuration structure (it's one level off), and we don't always want
to share it into containers into which we're sharing ~/co/, so just
get it out of the way. If we want cpd to be able to change into the
virtualenv dirs, we can add that configuration explicitly.

The compatibility code to unset the old $WORKON_HOME can be removed in
a while; it's needed only for existing shell processes trying to
update.