bin/watchrun: Command is now argument; watched files become options
The default command has been removed (it was a bit cjs-specific) and it's
now specified on the command line as non-option arguments, making it easier
to add various options to the command.

That means we now need to use option arguments to specify the
files/directories to watch, which we do with the `-f` option. But this is
less frequently needed now as if no `-f` arguments are given watchrun asks
Git what files it's tracking and uses that list. (It actually tracks
recursively the entire contents of directories at the top level, rather
than the individual tracked files below them. WARNING: this may track
object files if you put them in tracked directories! But that's bad
practice.)

We also now internally print separator lines between runs rather than
making the user specify a command that does this. This can be disabled with
the `-s` option.