bin/watchrun: $SHELL -i was definitely broken; instead use a Bash subshell
Running `$SHELL -i` was causing the shell to receive a SIGSTOP for whatever
reason, probably because it wasn't truly in an interactive environment.
At any rate, just use a `( ... )` subshell (which now forces the command
to be `bash`, since that's what this script's interpreter is) and manually
source ~/.bashrc before we start.

I'm not entirely clear that this is correct, but it's "more correct" than
what we were doing before.