bin/watchrun: Consume "extra" events _after_ running the command
Rather than waiting for more commands before we run, we now consume any
events after the first _after_ we run the command.

This has the minor benefit of not delaying the command run slightly, and
the major benefit of not re-triggering if any watched files/dirs were
modified during the run. In particular, the latter happens in __pycache__
directories containing the newly-compiled Python bytecode generated during
the run.