bin/watchrun: Ignore ^./.git/ paths
For whatever reason yet to be debugged, watchrun on an individual file
in the current directory doesn't work; we receive events only for
files in subdirectories (whether with or without --recursive). So in
situations where we're changing a file in the top level dir, we use
`-f .`. This use to pick up changes in .git/, e.g. from `git status`.
Since there's no obvious reason ever to want to rebuild on changes in
.git/, always exclude it.