.bashrc.inb7: End with `true`
If the last test of the file is P && Q and P is false, the shell will
be left with $? ≠ 0, and sourcing this file alone will exit with that
final error code. Since the last condition is tested, we don't actually
have an error condition after that even if the test fails; a `true`
at the end fixes this.