bin/watchrun: Run when untracked files are changed, too
The list we were getting from git-ls files included only files in the
index, which meant that changing a new files or one that's never committed
wouldn't trigger a run. We now incldue these as well.

This doesn't fix all the bugs; changing a top-level file still doesn't seem
to trigger a run, even though  though we're explicitly giving it to the
inotifywait (at least when it already exists). This may be related to using
the -d option to inotifywait.