bin/watchrun: Do recursive watches; ignore ./.build/
The previous understanding of how inotifywait worked was wrong; what was
missing is that it's watching _inodes_, not filenames. Rewrite the comments
extensively to make this clear and fix the watching problems we had by
watching recursively.

Also now ignore changes in ./.build/ since if they occurred without a
source file being changed we're probably not interested in rebuilding.