.bashrc.inb7: Partial fix for Docker container detection for Linux 5
This adds `:/system.slice/docker-` to the existing `:/docker/ as something
in /proc/1/cgroup that will identify this shell as containerized; it
matches `0::/system.slice/docker-#.scope`, where `#` is a long
crypto-string. That seems to be related to Linux 5: it came up after an
upgrade to Debian 11, moving from Linux 4.19 to 5.10. However, it works
only for containers that were built on the earlier system.

In new containers, /proc/1/cgroup contains just `0::/`, though on the host
the /proc/#/cgroup file for that process still contains
`0::/system.slice/docker-#.scope`. I am guessing that this is a change to
Docker (possibly supported by kernel changes) that allows it to better hide
the host's cgroup information from the containerized process.

The host on Linux 5 has `0::/init.scope` in /proc/1/cgroup; perhaps trying
to identify that one is on a host as opposed to in a container is a better
way to go. But that sounds like a systemd name, so we need to check some
non-systemd and some Linux 4.x systems to see how they behave.