.bashrc.inb7: Change docker prompt check to look for 'init' as proc 1
We check the current argv[0] from /proc/1/cmdline to see if it contains the
word 'init'; if not, we assume we're in a Docker container and prefix the
prompt with 'd'. This is actually a better heuristic anyway, since if we're
in a container running its own actual init system, we probably want to
treat it as a proper machine, and not a 'container' that has been forked
off to run something more limited.

This reverses the way we used to check things, which was to try to figure
out if we were in a container. That has several problems, some of which are
mentioned in the previous commit message.