.tmux.conf,dh/setup: Enable italics in tmux and tmate
When tmux or tmate starts, it doesn't do italics if the default
terminal type for inside the tmux/tmate is `screen*`. So here we
add a variant with a different name that does have italics so
that tmux/tmate will also display them, and interpret them correctly
when rewriting the codes for the end user's real terminal.

This information comes from the tmux and tmate FAQs:
• https://github.com/tmux/tmux/wiki/FAQ
• https://github.com/tmate-io/tmate/blob/master/FAQ