.tmux.conf: Pass title changes through to my window manager
This allows me to use `xt` in my tmate sessions and have it change my X11
window title. This seems to work only when hosting a session; when
connecting as a client using this config the message changes don't update
my status line.

We prefix the title with a big `TMATE:` because it seems helpful to
security to keep in mind that a particular window is shared.