.tmux.conf: Suffix should be prefix, and change to [TM]
I didn't actually move it to the prefix position in my last commit,
just thought it.

And this is actually used for tmux as well as tmate; that's fine for the
moment since a notification that we're using tmux in a window is probably
handy, too, given how rarely I do that.