.bashrc c/calc/vc: c is calc without need for quoting; improve vc
`c` is an alias that works essentially the same way as `vc`: interpret all
arguments to the alias as a comment so nothing need be quoted, and then
re-interpret the arguments from the history, quoting them to be passed to
the actual shell function.

We improved on the previous vc method by using `fc -n` so we need not
remove the line number and also improved the regular expression that
removes the alias from the front to work with all whitespace, not just
spaces. `vc` has been fixed to use this new method, too.