.bashrc.inb7: Always set $WORKON_HOME
Previously we set it only if we had virtualenvwrapper available in the
default (system or user) Python configuration. But we want to move towards
a default that has nothing (not even Pip); this is the first step towards
that. (programs in ~/.local/bin/ programs will later handle finding the
correct environments under $WORKON_HOME--or creating them--for whatever
they want to run.)