.bash_profile: export IGNOREEOF=2
This twice ignores Ctrl-D with no preceeding unprocessed input (printing
instead 'Use "exit" to leave the shell.') before actually exiting.

This is particularly useful when a Python virtualenv is activated and I
forget that I must use `deactivate`, not exit, to deactivate the
environment. Also in tmate sessions to avoid accidentaly exiting.