.ssh/known_hosts: Add `github.com` to existing `github.com.`
Enough people use e.g., `git@github.com:...` as default sources in
.gitmodule files and elsewhere that we should just accept this. Trying to
deal with it manually after (or even before) is a pain, and that just gets
worse when automated setup scripts are trying to do that for you.

So this is a hack, but so far I've not found a better solution.