bin/dropbox: Apply substitutions for all @…@ markers
This is normally done by the Debian build system. We use the Debian
values as of the current build, except that we just use an empty
string for the logo and icon image data as those are huge and hard to
read. (The program seems fine with them being empty.)