.gitconfig,.config/git/ignore: Global ignore file with `tags` in it
I'm a little worried that this might conflict with an actual
committed file in a repo one day, but I'm hoping that `tags`
is well known enough that nobody would ever use it. (I bet I'm
wrong about this though.)

Probably the proper solution is to use a more clear name (.ctags
or something even more likely to be unique), reconfigure my Vim
to do that, and add a Vim command to generate it at the
appropriate place or whatever.