.config/htop/: Rework meters
Individual CPU meters on one side, all the rest on the other, as this
works better on higher-cpu-count systems.

Tried with the inidvidual CPUs on the left, because that side is more
likely to be covered than the right, but that crammed it up against
the other side too hard becuase the individual meter bank appears to
be offset to the right.