quick-setup: Do not remove and rewrite .ssh/authorized_keys
If there's already a .ssh/authorized_keys on the host, that
may be configured that way for a reason. (E.g., it may be
using a work key rather than my personal key.) So we don't
want to overwrite that automatically with the personal key
from cjs1/dot/ssh/authorized_keys.inb4.

So now on systems without a .ssh/authorized_keys the personal
key will be installed; on systems with a .ssh/authorized_keys
a warning will be generated that .ssh/authorized_keys won't
be overwritten.

That warning is slightly annoying, but it's not clear what
the best way is to avoid that and and yet still install a
.ssh/authorized_keys with the personal key when it's not
present. This could be done with code in dh/setup, but then
we lose (or have to re-implment) the automatic detection of
external changes to that file.

So for the moment we just live with the warning, which is
actually something we kind of want on a regular basis
anyway, just maybe not _all_ the time.