.gitconfig: Clarify diff and color.diff settings (no settings changes)
diff.colorMoved has many options; use `default` (which is `zebra`
at the moment) to clarify that it's not just on/off.

In the colour section head comments, explain what `normal` and
`default` mean.

To color.diff add all of the color options (including lots of
new ones for diff.colorMoved), and sort them in a more sensible
order. (This is not the same order as the manpage.)