.gitconfig.inb4: Add diff.colors for func, old/newMoved
In the hunk header, the function name (diff.func) is now bold instead
of the default normal text, which makes it stand out a lot better.

diff.color.oldMoved and .newMoved are used for text that has been
moved from one part of the file to the other, as opposed to simple
deletions or brand new text. The former are now yellow and cyan, so as
to stand out less than red and blue for the latter.

Yes, this re-uses the same yellow we use for the diff.frag hunk
header, but the two are still easily distinguished. We also tried
magenta, but that stands out more than we would like for code that
hasn't actually been removed.