.bashrc dr(): Still use or own config even if we have to sudo to run docker
If we `sudo docker`, we may be running it with HOME=/root instead of
HOME=/home/cjs, so it will use the "wrong" $HOME/.docker/ for config.json
etc. This results in our access tokens not being used, and new access
tokens (from `dr login`) being put in /root/.docker/config.json.

More confusing is that we can run `docker login` (not `dr login`) and
successfully authenticate because it doesn't connect to the local Docker
daemon when doing this. So that update /home/cjs/.docker/config.json and
gives us a 'Login Successful' message, but when we `sudo docker` it
still wouldn't use that.