.bashrc.inb7: bind -f ~/.inputrc to refresh readline config
Note that this is executed only in interactive mode; doing it when
not in interactive mode can produce an error message.

No reason to have this as a separate step when sourcing .bashrc
can do it along with everything else.

Also update quick-setup message at end to remove the suggestion
to `bind -f ~/.inputrc`.