.bashrc __cjs_prompt_host: Re-order, add `f` for "foreign host"
If we're on a "foreign" host, defined as one whose `hostname -f`
does not match *.cynic.net, add an `f` to the prompt to remind us
to look at the title bar to see where we are. But the `c` in a
prompt indicating we're in a container overrides this, since all
containers are essentially "foreign" hosts.

The prompt chars are now re-ordered to put `t`/tmate first (it's
the most important), followed by screen, followed by host indicator.