.bash_profile: Allow suppressing fortune with $__0cjs_nofortune
We use this file now not only for our own logins, but to set up a
full cjs environment by sourcing it as another user. When we do
this we don't want to try to print a fortune because a) we don't
want to be that verbose, and b) we may not have access to
~cjs/.home/unfortunate/. So anybody sourcing this can now set the
$__0cjs_nofortune variable to any value and the fortune cookie
will be skipped.

Perhaps we want to do this for the terminal setting, too.

(This does seem a bit of a hack; perhaps we can improve it later.)