.tmux: Change default term to rxvt-unicode-256color
For some reason this provides better colours for my Vim configuration
than tmux-256color does, though I don't know why. (Something to do
with how it maps the 16 colours that Vim uses to a selection of the
256 colours the terminal supports?)

This also happens to drop the 'Terminal type?' prompt when starting up
because .bash_profile doesn't prompt to confirm the terminal if the
name matches `rxvt*`. This probably isn't ideal, since we want to be
clear on what terminal definition we're using when sharing with
others, but we'll fix this later if it becomes an issue.

We leave in the old tmux-256color setting as documentation, and
override it on the next line.