.stack/hooks/ghc-install.sh: Copy from ghcup distribution
This is the same hook that the ghcup[1] install script will install
when $BOOTSTRAP_HASKELL_MINIMAL is not set. It comes from [2] as of
the commit date.

We keep a copy in dot-home, rather than just letting 3pinst[3] install it,
because we regularly want to remove ~/.stack/ to ensure we're doing clean
builds, and that also removes the hook. (We still have to remember to do a
`dot-home-update` after the removal, though.)

[1]: https://www.haskell.org/ghcup/
[2]: https://www.haskell.org/ghcup/sh/hooks/stack/ghc-install.sh
[3]: https://github.com/0cjs/dot-home-cjs0/blob/main/bin/3pinst