bin/stack: Wrapper script to ensure our stack config is set up
We tend to `rm -rf ~/.stack/` when we want to clean everything and
re-download/build again, but this also removes our configuration file
and hook. This wrapper ensures that, before running stack, these are
installed if they don't already exist. (It also warns if they are
different from the versions in ~/.home/cjs1/.)

XXX This uses `realpath`, and so won't work on MacOS. (It needs
realpath or something similar to make dot-home-compatible links. We
need to figure out maybe how to call dot-home code to do this.)

This unfortunately introduces yet another dependncy on having fixed
names (in this case `cjs1`) for these modules under .home/; it's not
currently clear how to fix this. (Perhaps by looking at the path of
the symlink by which we were run?)

The .local/bin/ symlinks code in dh/setup now does nothing, but we
leave the code in place for reference should we need to do this kind
of thing again.