.stack/hooks/ghc-install: Fail earlier if we know we can't install
In containers we have a mostly read-only ~/.ghcup; we never install GHC
from within the container but instead must do the install from outside
the container. (This helps avoid malicious users within the container
installing a tainted GHC.)

So we check to see if ~/.ghcup/ghc/ is writable and, if not, fail right
away, letting Stack do its own install. The script previously always
attempted to do the install, which would do the download and unpack
before failing.