.bashrc: Fix test for `stack` existence now that we always have my wrapper
Previously we checked to see we could run `stack`. We now always can,
because the wrapper is always there, and it prints an error if it
can't find an actual `stack` binary. So now our test is to run `stack
--version` which will fail if the wrapper can't find a binary.

This gets rid of the spurious 'ERROR: stack executable not found'
message on systems where there isn't a proper `stack` binary.