.bashrc.inb7: Add $HOME/.nvm setup (if present)
NVM is the NodeJS installation/version manager.

This is unchanged from the code that the NVM installer[1] adds to
~/.bashrc. It always sets $NVM_DIR even if it's not present, but we
just leave that because it's not annoying enough that we want to
use a non-standard version.

The installer script has a bug: it assumes that the user's default
remote name is `origin` and fails if it is not. PR #3654 [2] has been
filed to fix this (it just adds `-o origin` to the clone command), but
is stalled for unknown reasons.

[1]: https://nodejs.org/en/download
[2]: https://github.com/nvm-sh/nvm/pull/3654