.config/git/ignore: /.claude/settings.local.json
This was silently added as `**/.claude/settings.local.json` by the
`claude` command-line program: smart that it knows to do it to my
personal ignore rather than the repo ignore!. And, while I hate that
it ignores in every dir (surely it always uses the root of the repo),
they were also nice enough to be explicit about that with `**/`!

Anyway, we drop the `**/` and we'll see how it works being rooted.