.gitconfig.inb4: advice.skippedCherryPicks = false
Suppress the following advice:

    hint: use --reapply-cherry-picks to include skipped commits
    hint: Disable this message with "git config advice.skippedCherryPicks false"

Generally, I always want to skip commits in my branch (usually a dev
branch) that are on the rebase target (usually main@{upstream}). The hint
to look at --reapply-cherry-picks is actually kind of handy, but now that I
know there are various options for `git rebase` related to that, the
"warning: skipped previously applied commit c4e21fd4" message will remind
me that I can look those up if I need to.

As an aside, `git rebase --empty=ask` option looked like it could be
useful, but it's not just as "yes/no" thing but drops you at a prompt in
the middle of the rebase suggesting you use `git rebase --skip` or
whatever, so not something I'd want to enable by default.